Improved consistency of Ball physics. Gave it a material for contrast.
Changed the Ball's physics to have it's velocity's magnitude kept constant all the time. It was also given the BlankWhite material to give it better contrast with the floor.

Also realized that I should have included the "Saved" folder and a few others in the gitignore. I'll do so in future UE4 projects.